+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQGZE4FZB6BTK5MJT8BWXRT0"
+        "01KQH4564FD74X2DWJ5H3Z05YM"
     )
     r"""The unique identifier of the data source"""