+++ packages/orq-rc/RELEASES.md
@@ -7469,3 +7469,13 @@ Based on:
 - [python v4.9.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.14 - packages/orq-rc
+
+## 2026-05-01 06:41:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.15] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.15 - packages/orq-rc